Official release hotfix.

+++ WalkOfLife/Framework/Events/GameLoop/UpdateTicked/SuperModeActivationTimerUpdateTickedEvent.cs
@@ -7,7 +7,7 @@ namespace TheLion.Stardew.Professions.Framework.Events
 	{
 		private const int BASE_SUPERMODE_ACTIVATION_DELAY = 60;
 
-		private int _superModeActivationTimer = BASE_SUPERMODE_ACTIVATION_DELAY * ModEntry.Config.SuperModeActivationDelay;
+		private int _superModeActivationTimer = (int)(BASE_SUPERMODE_ACTIVATION_DELAY * ModEntry.Config.SuperModeActivationDelay);
 
 		/// <inheritdoc/>
 		public override void OnUpdateTicked(object sender, UpdateTickedEventArgs e)
